Revert "Bug 2082667: Separate controller for the node draining"
