+++ pkg/controller/machine/controller_test.go
@@ -289,38 +289,6 @@ func TestReconcileRequest(t *testing.T) {
 			},
 		},
 	}
-	machineDeletingAlreadyDrained := machinev1.Machine{
-		TypeMeta: metav1.TypeMeta{
-			Kind: "Machine",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:              "delete-drained",
-			Namespace:         "default",
-			Finalizers:        []string{machinev1.MachineFinalizer, metav1.FinalizerDeleteDependents},
-			DeletionTimestamp: &time,
-			Labels: map[string]string{
-				machinev1.MachineClusterIDLabel: "testcluster",
-			},
-		},
-		Spec: machinev1.MachineSpec{
-			LifecycleHooks: machinev1.LifecycleHooks{
-				PreDrain: []machinev1.LifecycleHook{
-					{
-						Name:  "protect-from-drain",
-						Owner: "machine-api-tests",
-					},
-				},
-			},
-			ProviderSpec: machinev1.ProviderSpec{
-				Value: &runtime.RawExtension{
-					Raw: []byte("{}"),
-				},
-			},
-		},
-		Status: machinev1.MachineStatus{
-			Conditions: machinev1.Conditions{*conditions.TrueCondition(machinev1.MachineDrained)},
-		},
-	}
 
 	type expected struct {
 		createCallCount int64
@@ -380,9 +348,9 @@ func TestReconcileRequest(t *testing.T) {
 			existsValue: false,
 			expected: expected{
 				createCallCount: 0,
-				existCallCount:  0,
+				existCallCount:  1,
 				updateCallCount: 0,
-				deleteCallCount: 0,
+				deleteCallCount: 1,
 				result:          reconcile.Result{},
 				error:           false,
 				phase:           phaseDeleting,
@@ -393,10 +361,10 @@ func TestReconcileRequest(t *testing.T) {
 			existsValue: true,
 			expected: expected{
 				createCallCount: 0,
-				existCallCount:  0,
+				existCallCount:  1,
 				updateCallCount: 0,
-				deleteCallCount: 0,
-				result:          reconcile.Result{},
+				deleteCallCount: 1,
+				result:          reconcile.Result{RequeueAfter: requeueAfter},
 				error:           false,
 				phase:           phaseDeleting,
 			},
@@ -419,10 +387,10 @@ func TestReconcileRequest(t *testing.T) {
 			existsValue: true,
 			expected: expected{
 				createCallCount: 0,
-				existCallCount:  0,
+				existCallCount:  1,
 				updateCallCount: 0,
-				deleteCallCount: 0,
-				result:          reconcile.Result{},
+				deleteCallCount: 1,
+				result:          reconcile.Result{RequeueAfter: requeueAfter},
 				error:           false,
 				phase:           phaseDeleting,
 			},
@@ -435,32 +403,6 @@ func TestReconcileRequest(t *testing.T) {
 				existCallCount:  0,
 				updateCallCount: 0,
 				deleteCallCount: 0,
-				result:          reconcile.Result{},
-				error:           false,
-				phase:           phaseDeleting,
-			},
-		},
-		{
-			request:     reconcile.Request{NamespacedName: types.NamespacedName{Name: machineDeletingAlreadyDrained.Name, Namespace: machineDeletingAlreadyDrained.Namespace}},
-			existsValue: true,
-			expected: expected{
-				createCallCount: 0,
-				existCallCount:  1,
-				updateCallCount: 0,
-				deleteCallCount: 1,
-				result:          reconcile.Result{RequeueAfter: requeueAfter},
-				error:           false,
-				phase:           phaseDeleting,
-			},
-		},
-		{
-			request:     reconcile.Request{NamespacedName: types.NamespacedName{Name: machineDeletingAlreadyDrained.Name, Namespace: machineDeletingAlreadyDrained.Namespace}},
-			existsValue: false,
-			expected: expected{
-				createCallCount: 0,
-				existCallCount:  1,
-				updateCallCount: 0,
-				deleteCallCount: 1,
 				result:          reconcile.Result{},
 				error:           false,
 				phase:           phaseDeleting,
@@ -510,7 +452,6 @@ func TestReconcileRequest(t *testing.T) {
 					&machineDeletingPreTerminateHook,
 					&machineFailed,
 					&machineRunning,
-					&machineDeletingAlreadyDrained,
 				),
 				scheme:   scheme.Scheme,
 				actuator: act,
